add implementacao final

+++ src/main/java/com/fiap/tech/pedidos/domain/usecase/DeletePedidoUseCase.java
@@ -1,9 +1,9 @@
 package com.fiap.tech.pedidos.domain.usecase;
 
+import com.fiap.tech.pedidos.domain.exception.ResourceNotFoundException;
 import com.fiap.tech.pedidos.infra.repository.PedidoRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @AllArgsConstructor
@@ -11,9 +11,13 @@ public class DeletePedidoUseCase {
 
 	private final PedidoRepository pedidoRepository;
 
-	@Transactional
-	public void deletarPedido(Long id) {
-		pedidoRepository.deleteById(id);
+	public void deletarPedido(Long id) throws ResourceNotFoundException {
+//pedidoRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException(String.format("O pedido %d já não existe", id)));
+		if(pedidoRepository.findById(id).isPresent()) {
+			pedidoRepository.deleteById(id);
+		}else{
+			throw new ResourceNotFoundException(String.format("O pedido %d já não existe", id);
+		}
 	}
 
 }
